Make consumers able to recover from invalid state (#6)
Small bugfixes:

Add periodical binding check for consumers to make consumers able to detect if the binding, queue or exchange was deleted/disappeared and make it able to recover.
It can happen the forwarder actor was not termination is not finished (shutdown) while during
the channel creation ChannelOwner trying to create new forwarder. This causing:
akka.actor.InvalidActorNameException: actor name [forwarder] is not unique! (This causing additional restart of ChannelOwner actor)
Asana task:
https://app.asana.com/0/0/869432835856947/f